requirements.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
pydantic>=2.11,<3.0
pandas>=2.2,<3.0
numpy>=2.2,<3.0
scikit-learn>=1.7,<2.0
streamlit>=1.45,<2.0
httpx>=0.28,<1.0
paho-mqtt>=2.1,<3.0
azure-storage-blob>=12.26,<13.0
pytest>=8.0,<10.0
